feat: integrate beta labs toggle for vector basemap in navbar actions menu
