+++ src/components/IntegratedToolbar/IntegratedToolbar.tsx
@@ -6,7 +6,6 @@ import {
   CloudHail, 
   LayoutGrid,
   Layers,
-  SlidersHorizontal,
   Calendar,
   ChevronLeft,
   ChevronRight,
@@ -22,13 +21,6 @@ import {
   Redo2,
 } from 'lucide-react';
 import { Button } from '../ui/button';
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuLabel,
-  DropdownMenuTrigger,
-} from '../ui/dropdown-menu';
 import {
   Tooltip,
   TooltipContent,
@@ -60,7 +52,6 @@ import {
   redoLastEdit,
   undoLastEdit,
 } from '../../store/forecastSlice';
-import { setFeatureFlag } from '../../store/featureFlagsSlice';
 import { toggleGhostOutlook } from '../../store/overlaysSlice';
 import { ForecastMapHandle } from '../Map/ForecastMap';
 import CycleHistoryModal from '../CycleManager/CycleHistoryModal';
@@ -180,9 +171,6 @@ interface IntegratedToolbarViewProps {
   probabilityHandlers: Record<string, () => void>;
   currentColor: string;
   isLowProb: boolean;
-  showBetaLabs: boolean;
-  vectorBasemapEnabled: boolean;
-  onToggleVectorBasemap: (enabled: boolean) => void;
 }
 
 /** Toolbar actions panel: Save, Load, Export, Package download, Cycle History, Copy from Previous, and Reset buttons. */
@@ -289,64 +277,6 @@ const ToolbarToolsSection: React.FC<{
         className="h-14 w-14 lg:h-16 lg:w-16 bg-red-500/20 hover:bg-red-500/30 border-red-500/50 text-red-700 dark:!bg-red-500/20 dark:hover:!bg-red-500/30 dark:border-red-500/50 dark:text-red-400"
         onClick={onOpenResetConfirm}
       />
-    </div>
-  </div>
-);
-
-/** Trigger button for the beta Labs dropdown; forwards Radix trigger props/ref to the underlying button. */
-const BetaLabsTriggerButton = React.forwardRef<
-  HTMLButtonElement,
-  React.ComponentPropsWithoutRef<typeof Button>
->(({ className, ...props }, ref) => (
-  <Button
-    ref={ref}
-    variant="outline"
-    className={cn(
-      'h-14 lg:h-16 min-w-[128px] justify-start gap-2 border-sky-500/50 bg-sky-500/10 text-sky-700 hover:bg-sky-500/20 dark:!bg-sky-500/15 dark:hover:!bg-sky-500/25 dark:border-sky-500/50 dark:text-sky-300',
-      className
-    )}
-    {...props}
-  >
-    <SlidersHorizontal className="h-4 w-4" />
-    <span>Labs</span>
-  </Button>
-));
-
-BetaLabsTriggerButton.displayName = 'BetaLabsTriggerButton';
-
-/** Beta-only Labs menu content for quickly comparing legacy and experimental map rendering. */
-const BetaLabsMenuContent: React.FC<{
-  vectorBasemapEnabled: boolean;
-  onToggleVectorBasemap: (enabled: boolean) => void;
-}> = ({ vectorBasemapEnabled, onToggleVectorBasemap }) => (
-  <DropdownMenuContent align="start" className="w-72">
-    <DropdownMenuLabel>Beta map experiments</DropdownMenuLabel>
-    <DropdownMenuCheckboxItem
-      checked={vectorBasemapEnabled}
-      onCheckedChange={(checked) => onToggleVectorBasemap(Boolean(checked))}
-    >
-      Vector basemap + opaque outlook fills
-    </DropdownMenuCheckboxItem>
-  </DropdownMenuContent>
-);
-
-/** Quick beta-only experiment menu so testers can compare old and new forecast map rendering. */
-const BetaLabsSection: React.FC<{
-  vectorBasemapEnabled: boolean;
-  onToggleVectorBasemap: (enabled: boolean) => void;
-}> = ({ vectorBasemapEnabled, onToggleVectorBasemap }) => (
-  <div className="border-r border-border pr-2 lg:pr-4">
-    <div className="flex flex-col gap-3">
-      <label className="text-xs font-medium text-muted-foreground uppercase tracking-wide">Beta Labs</label>
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <BetaLabsTriggerButton />
-        </DropdownMenuTrigger>
-        <BetaLabsMenuContent
-          vectorBasemapEnabled={vectorBasemapEnabled}
-          onToggleVectorBasemap={onToggleVectorBasemap}
-        />
-      </DropdownMenu>
     </div>
   </div>
 );
@@ -741,7 +671,6 @@ const IntegratedToolbarView: React.FC<IntegratedToolbarViewProps> = (props) => {
     lowProbabilityOutlooks, visibleGhostOutlooks, outlookTypeHandlers, ghostOutlookHandlers,
     probabilities, probabilityHandlers,
     currentColor, isLowProb,
-    showBetaLabs, vectorBasemapEnabled, onToggleVectorBasemap,
   } = props;
 
   const ghostTypes = availableTypes.filter((type) => type !== activeOutlookType);
@@ -759,12 +688,6 @@ const IntegratedToolbarView: React.FC<IntegratedToolbarViewProps> = (props) => {
               isSaved={isSaved} canUndo={canUndo} canRedo={canRedo}
               isExporting={isExporting} isPackageDownloading={isPackageDownloading}
             />
-            {showBetaLabs ? (
-              <BetaLabsSection
-                vectorBasemapEnabled={vectorBasemapEnabled}
-                onToggleVectorBasemap={onToggleVectorBasemap}
-              />
-            ) : null}
             <ToolbarForecastDaySection
               isEditingDate={isEditingDate} tempDate={tempDate} cycleDate={cycleDate}
               currentDay={currentDay} days={days}
@@ -1057,12 +980,7 @@ export const IntegratedToolbar: React.FC<IntegratedToolbarProps> = ({
   cloudTools = null,
 }) => {
   const dispatch = useDispatch();
-  const vectorBasemapEnabled = useSelector((state: RootState) => state.featureFlags.vectorBasemapEnabled);
   const model = useToolbarDataModel(mapRef, addToast);
-  const showBetaLabs = __GFC_BETA_MODE__;
-  const handleToggleVectorBasemap = useCallback((enabled: boolean) => {
-    dispatch(setFeatureFlag({ feature: 'vectorBasemapEnabled', enabled }));
-  }, [dispatch]);
 
   const localUi = useToolbarLocalUi(model.cycleDate);
   const handlers = useToolbarActionHandlers({
@@ -1094,9 +1012,6 @@ export const IntegratedToolbar: React.FC<IntegratedToolbarProps> = ({
       onCancelReset={localUi.handleCancelReset}
       onTempDateChange={localUi.handleTempDateChange}
       onStartDateEdit={localUi.handleStartDateEdit}
-      showBetaLabs={showBetaLabs}
-      vectorBasemapEnabled={vectorBasemapEnabled}
-      onToggleVectorBasemap={handleToggleVectorBasemap}
       {...handlers}
       {...localUi}
       {...model}
